Implement external pose functionality
It's not possible to access the habitat-sim from more than one thread, so the
agent's pose can't be set from a ROS callback. Thus the ROS callback just sets a
variable and the agent pose is periodically set to that. The whole node is now
implemented within a single class to make sharing state between methods easier.