Fix the received pose and observation having different timestamps
The timestamp of published images was the time after they had been
rendered instead of the timestamp of the received pose. This caused them
to occasionally be out of sync.

Merged the teleport and render functions to receive the pose and its
timestamp in the same lock to prevent them being out of sync.